(PUP-9258) Refactor loading of PAL
This refactors PAL (without any need for users to change anything):

* the implementation is now in multiple files under `puppet/pal`
* the `puppet_pal` file now just requires puppet and the pal api
* puppet now also requires the PAL api

While it would be possible to drop the loading of `puppet/pal/pal_api`
because `puppet` requires it the longer term intention is that `pal`
will not require everything that puppet requires.
