+++ lib/puppet.rb
@@ -297,3 +297,4 @@ require 'puppet/util/storage'
 require 'puppet/status'
 require 'puppet/file_bucket/file'
 require 'puppet/plugins/configuration'
+require 'puppet/pal/pal_api'
